backend/requirements.txt
Django==4.2
user-agents==2.2.0
whitenoise==5.2.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
dj-database-url==0.5.0
drf-yasg==1.21.4
drf-spectacular==0.16.0
django-oauth-toolkit==2.2.0
django-filter==22.1
social-auth-app-django==4.0.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.0.0
djangorestframework==3.14.0
djoser==2.0.5
celery==5.0.5
drf-standardized-errors==0.12.5

# django-auth-ldap

